# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /watchlist:
    get:
      operationId: listWatchlist
      tags: [watchlist]
      summary: List my saved watchlist items
      description: >-
        Returns the caller's saved shows by default, each annotated with the
        names of other members of the given group who have rated or saved the
        same show. Pass userId to view another member's saved shows (requires a
        shared group).
      parameters:
        - name: userId
          in: query
          required: false
          description: View this member's saved shows instead of your own. Requires a shared group.
          schema:
            type: string
        - name: groupId
          in: query
          required: false
          description: Scope the "also engaged by" names to members of this group.
          schema:
            type: integer
      responses:
        "200":
          description: A list of saved watchlist items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WatchlistItem"
    post:
      operationId: createWatchlistItem
      tags: [watchlist]
      summary: Save a show to my watchlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WatchlistInput"
      responses:
        "201":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchlistItem"
  /watchlist/{id}:
    delete:
      operationId: deleteWatchlistItem
      tags: [watchlist]
      summary: Remove a show from my watchlist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    MediaType:
      type: string
      enum: [movie, tv]
...
    WatchlistItem:
      type: object
      required:
        - id
        - title
        - mediaType
        - createdAt
        - alsoEngagedBy
      properties:
        id:
          type: integer
        title:
          type: string
        mediaType:
          $ref: "#/components/schemas/MediaType"
        createdAt:
          type: string
          format: date-time
        alsoEngagedBy:
          type: array
          description: Display names of other group members who rated or saved this show.
          items:
            type: string
    WatchlistInput:
      type: object
      required:
        - title
        - mediaType
      properties:
        title:
          type: string
          minLength: 1
        mediaType:
          $ref: "#/components/schemas/MediaType"
...